requirements.txt
streamlit~=1.51.0
pillow~=12.0.0